refactor: extract IndustrySelect component; rearrange component graph
