+++ src/shared/constants.ts
@@ -49,6 +49,7 @@ export const companyEvents = {
   COMPANY_ADDRESSES: 'company/addresses',
   COMPANY_ADDRESSE_EDIT: 'company/address/edit',
   COMPANY_INDUSTRY: 'company/industry',
+  COMPANY_INDUSTRY_SELECTED: 'company/industry/selected',
   COMPANY_FEDERAL_TAXES_UPDATED: 'company/federalTaxes/updated',
 } as const
